Fix estimate listing indirect dependencies as direct
As noted in 7bc3609: "sometimes go mod graph returns a little bit too much
direct dependencies for an unknown reason".

This is not anymore an unknown reason, as it is in fact by design, even
if it is a bit strange (see https://github.com/golang/go/issues/27309).

We do not want estimate to show indirect dependencies as direct ones, so
we filter them out. To do that, we first need to get the list of the
direct dependencies of the module we want to package. But as we had to
create a dummy go module to continue using "go get" (which is a vcs
agnostic way to download a Go module), we first need to find in the
GOPATH the directory that contains the module we want to package.